requirements.txt
# Telegram Bot
aiogram>=3.4.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Configuration
python-dotenv>=1.0.0

# Validation
pydantic>=2.0.0

# Debug (стиль Глеба)
icecream>=2.1.0